refactor(agenda): split resolver facade
Move resolver quality, Legistar policy, HTML loading, page enrichment, and source-priority orchestration into focused modules.

Keep pipeline.agenda_resolver as the public patch surface so existing callers and tests keep importing the same names.
